Move scoped validation requests to a module under planning applications
