+++ engines/bops_api/config/routes.rb
@@ -18,7 +18,9 @@ BopsApi::Engine.routes.draw do
         get :submission, on: :member
         get :search, on: :collection
 
-        resources :validation_requests, only: [:index]
+        scope module: "planning_applications" do
+          resources :validation_requests, only: [:index]
+        end
       end
 
       namespace :public do
